rm `fuse::util`
