+++ query/src/datasources/table/fuse/io/location_gen.rs
