+++ query/src/datasources/table/fuse/io/mod.rs
@@ -13,17 +13,17 @@
 // limitations under the License.
 //
 
-pub(crate) use block_appender::*;
+pub use block_appender::*;
 pub use block_reader::*;
-pub use segment_reader::*;
-
-// consider remove these, read_util seems to be enough (type could be inferred)
-mod segment_reader;
-// end
+pub use col_encoding::*;
+pub use location_gen::*;
+pub use meta_info_reader::*;
 
 mod block_appender;
 mod block_reader;
-pub(crate) mod meta_info_reader;
+mod col_encoding;
+mod location_gen;
+mod meta_info_reader;
 
 #[cfg(test)]
 mod block_appender_test;
