+++ query/src/datasources/table/fuse/statistics/util_test.rs
@@ -22,14 +22,16 @@ use common_datavalues::DataType;
 use common_datavalues::DataValue;
 use common_exception::ErrorCode;
 
-use super::statistic_helper;
+use crate::datasources::table::fuse::statistics::block_meta_accumulator::BlockMetaAccumulator;
+use crate::datasources::table::fuse::statistics::stats_accumulator::StatisticsAccumulator;
+use crate::datasources::table::fuse::statistics::util;
 use crate::datasources::table::fuse::table_test_fixture::TestFixture;
 
 #[test]
 fn test_ft_stats_block_stats() -> common_exception::Result<()> {
     let schema = DataSchemaRefExt::create(vec![DataField::new("a", DataType::Int32, false)]);
     let block = DataBlock::create_by_array(schema, vec![Series::new(vec![1, 2, 3])]);
-    let r = statistic_helper::block_stats(&block)?;
+    let r = util::block_stats(&block)?;
     assert_eq!(1, r.len());
     let col_stats = r.get(&0).unwrap();
     assert_eq!(col_stats.min, DataValue::Int32(Some(1)));
@@ -43,9 +45,9 @@ fn test_ft_stats_col_stats_reduce() -> common_exception::Result<()> {
     let schema = DataSchemaRefExt::create(vec![DataField::new("a", DataType::Int32, false)]);
     let col_stats = blocks
         .iter()
-        .map(statistic_helper::block_stats)
+        .map(util::block_stats)
         .collect::<common_exception::Result<Vec<_>>>()?;
-    let r = statistic_helper::column_stats_reduce_with_schema(&col_stats, &schema);
+    let r = util::column_stats_reduce_with_schema(&col_stats, &schema);
     assert!(r.is_ok());
     let r = r.unwrap();
     assert_eq!(1, r.len());
@@ -58,8 +60,8 @@ fn test_ft_stats_col_stats_reduce() -> common_exception::Result<()> {
 #[test]
 fn test_ft_stats_accumulator() -> common_exception::Result<()> {
     let blocks = TestFixture::gen_block_stream(10);
-    let mut stats_acc = statistic_helper::StatisticsAccumulator::new();
-    let mut meta_acc = statistic_helper::BlockMetaAccumulator::new();
+    let mut stats_acc = StatisticsAccumulator::new();
+    let mut meta_acc = BlockMetaAccumulator::new();
     blocks.iter().try_for_each(|item| {
         stats_acc.acc(item)?;
         meta_acc.acc(1, "".to_owned(), &mut stats_acc);
